Move timeouts mod to index-scheduler
