+++ crates/meilisearch/src/proxy/enterprise_edition.rs
@@ -26,31 +26,6 @@ use crate::error::MeilisearchHttpError;
 use crate::proxy::{Body, Endpoint, ProxyError, ReqwestErrorWithoutUrl};
 use crate::routes::SummarizedTaskView;
 
-mod timeouts {
-    use std::sync::LazyLock;
-
-    pub static CONNECT_SECONDS: LazyLock<u64> =
-        LazyLock::new(|| fetch_or_default("MEILI_EXPERIMENTAL_PROXY_CONNECT_TIMEOUT_SECONDS", 3));
-
-    pub static BACKOFF_SECONDS: LazyLock<u64> =
-        LazyLock::new(|| fetch_or_default("MEILI_EXPERIMENTAL_PROXY_BACKOFF_TIMEOUT_SECONDS", 25));
-
-    pub static REQUEST_SECONDS: LazyLock<u64> =
-        LazyLock::new(|| fetch_or_default("MEILI_EXPERIMENTAL_PROXY_REQUEST_TIMEOUT_SECONDS", 30));
-
-    fn fetch_or_default(key: &str, default: u64) -> u64 {
-        match std::env::var(key) {
-            Ok(timeout) => timeout.parse().unwrap_or_else(|_| {
-                panic!("`{key}` environment variable is not parseable as an integer: {timeout}")
-            }),
-            Err(std::env::VarError::NotPresent) => default,
-            Err(std::env::VarError::NotUnicode(_)) => {
-                panic!("`{key}` environment variable is not set to a integer")
-            }
-        }
-    }
-}
-
 impl<T, F> Body<T, F>
 where
     T: serde::Serialize,
@@ -253,7 +228,7 @@ where
 
                     let backoff = backoff::ExponentialBackoffBuilder::new()
                         .with_max_elapsed_time(Some(std::time::Duration::from_secs(
-                            *timeouts::BACKOFF_SECONDS,
+                            *index_scheduler::timeouts::BACKOFF_SECONDS,
                         )))
                         .build();
 
@@ -370,7 +345,9 @@ where
     let api_key = remote.write_api_key.clone();
 
     let backoff = backoff::ExponentialBackoffBuilder::new()
-        .with_max_elapsed_time(Some(std::time::Duration::from_secs(*timeouts::BACKOFF_SECONDS)))
+        .with_max_elapsed_time(Some(std::time::Duration::from_secs(
+            *index_scheduler::timeouts::BACKOFF_SECONDS,
+        )))
         .build();
 
     backoff::future::retry(backoff, move || {
@@ -384,8 +361,9 @@ where
 
         async move {
             let request = client.request(method, url).prepare(|request| {
-                let request =
-                    request.timeout(std::time::Duration::from_secs(*timeouts::REQUEST_SECONDS));
+                let request = request.timeout(std::time::Duration::from_secs(
+                    *index_scheduler::timeouts::REQUEST_SECONDS,
+                ));
                 let request = if let Some(body) = body { request.body(body) } else { request };
                 let request = if let Some(api_key) = api_key {
                     request.bearer_auth(api_key)
@@ -473,7 +451,8 @@ async fn try_proxy(
     body: Option<Bytes>,
 ) -> Result<SummarizedTaskView, backoff::Error<ProxyError>> {
     let request = client.request(method, url).prepare(|request| {
-        let request = request.timeout(std::time::Duration::from_secs(*timeouts::REQUEST_SECONDS));
+        let request = request
+            .timeout(std::time::Duration::from_secs(*index_scheduler::timeouts::REQUEST_SECONDS));
         let request = if let Some(body) = body { request.body(body) } else { request };
         let request =
             if let Some(api_key) = api_key { request.bearer_auth(api_key) } else { request };
